Clean exit (exit status==0) when vde commands receive SIGTERM

+++ vde-2/src/vde_autolink.c
@@ -186,7 +186,10 @@ static void sig_handler(int sig)
 	/*fprintf(stderr,"Caught signal %d, cleaning up and exiting", sig);*/
 	cleanup();
 	signal(sig, SIG_DFL);
-	kill(getpid(), sig);
+	if (sig == SIGTERM)
+		_exit(0);
+	else
+		kill(getpid(), sig);
 } 
 
 struct autolink *find_alink_pid(int pid);
