+++ vde-2/src/vde_pcapplug.c
@@ -96,7 +96,10 @@ static void sig_handler(int sig)
 {
 	cleanup();
 	signal(sig, SIG_DFL);
-	kill(getpid(), sig);
+	if (sig == SIGTERM)
+		_exit(0);
+	else
+		kill(getpid(), sig);
 }
 
 static void setsighandlers()
